test: Refac organizations test to include tests for Lat and Lon
